+++ src/axom/quest/tests/quest_mesh_view_util.cpp
@@ -17,7 +17,6 @@
   #include "axom/core.hpp"
   #include "axom/slic.hpp"
   #include "axom/primal.hpp"
-  #include "axom/core/ArrayIndexer.hpp"
   #include "axom/quest/MeshViewUtil.hpp"
   #include "axom/core/Types.hpp"